is-it-down/requirements.txt
opentelemetry-api==1.7.1
opentelemetry-exporter-otlp==1.7.1
opentelemetry-instrumentation-flask==0.26b1
opentelemetry-instrumentation-requests==0.26b1
opentelemetry-sdk==1.7.1
requests==2.26.0
opentelemetry-propagator-jaeger==1.7.1